+++ NEWS.md
@@ -3,6 +3,7 @@
 ## v1.2.0
 
 - Add -R --replace_metadata option. Default behavior is to merge keys.
+- Add -D --delete_keys option.
 
 ## v1.1.10